Update DTOs and services for long IDs and DI changes
Changed PlanogramId and ReplyNoteId to long in NewNoteDto. Updated PlanmShelfInfo and PlanmShelfInfoList to use nullable long/int for relevant IDs. Renamed properties in EditPlanApiController for consistency and added explicit long casts. Enabled DI for IClusterShelfRepository and IClusterPartRepository in PMServiceExtenstions.

+++ Dtos/Notes/NewNoteDto.cs
@@ -4,8 +4,8 @@
     {
 
         public string Note { get; set; }
-        public int PlanogramId { get; set; }
-        public int ReplyNoteId { get; set; }
+        public long PlanogramId { get; set; }
+        public long ReplyNoteId { get; set; }
 
 
     }
